# Receiver/openapi.yml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: localhost
    url: http://localhost:8080
info:
  description: This API receives reading events from a Bike Trainer and Heart Rate Monitor devices
  version: "1.0.0"
  title: Ride Data Statistics
  contact:
    email: akuanysheva@my.bcit.ca
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
tags:
  - name: devices
    description: Ride Data
paths:
  /ride:
    post:
      tags:
        - Ride data readings
      summary: Reports the statistics as a user is riding the application
      operationId: app.ride
      description: Adds ride data to the system
      responses:
        '201':
          description: item created
        '400':
          description: 'invalid input, object invalid'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/rideEvent'
        description: Inventory item to add
  /heartrate:
    post:
        tags:
          - heart rate readings
        summary: Reports a heart rate statistics
        operationId: app.heartrate
        description: Adds a heart rate readings to the system
        responses:
          '201':
            description: item created
          '400':
            description: 'invalid input, object invalid'
        requestBody:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/hrEvent'
          description: Inventory item to add
          
components:
  schemas:
    rideEvent:
      required:
        - ride_id
        - user_id
        - movie
        - timestamp
        - avg_speed #km/h
        - avg_power #watts/h
        - distance #km
      properties:
        ride_id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        user_id:
          type: string
          example: "A01258780"
        movie:
          type: string
          example: "John Wick 3"
        timestamp:
          type: string
          format: date-time
          example: '2016-08-29T09:12:33.001Z'
        avg_speed:
          type: integer
          example: 20
        avg_power: 
          type: integer
          example: 180
        distance: 
          type: integer
          example: 13
    hrEvent:
      type: object
      required:
        - ride_id # relates to ride data  event
        - user_id # relates to ride data event
        - device_id # unique
        - heart_rate # unique
        - max_hr # unique (at specific time)
        - min_hr # unique (at specific time)
        - timestamp #relates to ride data  event
      properties:
        ride_id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        user_id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        device_id:
          type: string
          example: WSD023
        heart_rate:
          type: integer
          example: 65
        max_hr:
          type: integer
          example: 100
        min_hr:
          type: integer
          example: 40
        timestamp:
          type: string
          format: date-time
          example: '2016-08-29T09:12:33.001Z'